break up loop.ts
