+++ src/compiler/index.ts
@@ -13,7 +13,6 @@ export * from "./identifier";
 export * from "./if";
 export * from "./indexed";
 export * from "./literal";
-export * from "./loop";
 export * from "./module";
 export * from "./namespace";
 export * from "./new";
@@ -29,3 +28,11 @@ export * from "./try";
 export * from "./unary";
 export * from "./variable";
 export * from "./yield";
+
+export * from "./break";
+export * from "./continue";
+export * from "./do";
+export * from "./for";
+export * from "./forOf";
+export * from "./loop";
+export * from "./while";
